Perovskite dataset
Combined 829 ABX3 and 624 A2BB'X6 perovskite structures all extracted from the Materials Project database. A quick overview of data segmentation of the different anions used for the learning exercise reveals; 65% as oxide-perovskites, 11% as fluoride-perovskites, 8% as chloride-perovskites, 5% as sulfide-perovskites, 4% as bromide-perovskites, 3% as selenide-perovskites, 2% as iodide-perovskites, and others. Infinite and finite bandgap values were also included in the pool of samples.